added testing and security to routes

+++ routes/api-routes.js
@@ -57,43 +57,57 @@ module.exports = function(app) {
 
   // CHECK FOR SECURITY. Post route to create a new task
   app.post("/api/create", (req, res) => {
-    console.log(req.user);
-    db.Task.create({
-      task: req.body.task,
-      due_date: req.body.due_date,
-      due_date_time: req.body.due_date_time,
-      category: req.body.category,
-      UserId: req.body.UserId
-    })
-      .then(newTask => res.status(201).json(newTask))
-      .catch(err => res.status(404).send(err));
-  });
-
-  app.put("/api/edit", (req, res) => {
-    db.Task.update(
-      {
+    if (!req.user) {
+      // The user is not logged in, send back a 403 status code
+      res.status(403).end();
+    } else {
+      db.Task.create({
         task: req.body.task,
         due_date: req.body.due_date,
         due_date_time: req.body.due_date_time,
-        category: req.body.category
-      },
-      {
-        where: {
-          id: req.body.id
+        category: req.body.category,
+        UserId: req.body.UserId
+      })
+        .then(newTask => res.status(201).json(newTask))
+        .catch(err => res.status(404).send(err));
+    }
+  });
+
+  app.put("/api/edit", (req, res) => {
+    if (!req.user) {
+      // The user is not logged in, send back a 403 status code
+      res.status(403).end();
+    } else {
+      db.Task.update(
+        {
+          task: req.body.task,
+          due_date: req.body.due_date,
+          due_date_time: req.body.due_date_time,
+          category: req.body.category
+        },
+        {
+          where: {
+            id: req.body.id
+          }
         }
-      }
-    )
-      .then(editedTask => res.status(200).json(editedTask))
-      .catch(err => res.status(404).send(err));
+      )
+        .then(() => res.status(200).send("Successfull update"))
+        .catch(err => res.status(404).send(err));
+    }
   });
 
   app.delete("/api/delete-task", (req, res) => {
-    db.Task.destroy({
-      where: {
-        id: req.body.id
-      }
-    })
-      .then(() => res.status(200).send("Task deleted"))
-      .catch(err => res.status(404).send(err));
+    if (!req.user) {
+      // The user is not logged in, send back a 403 status code
+      res.status(403).end();
+    } else {
+      db.Task.destroy({
+        where: {
+          id: req.body.id
+        }
+      })
+        .then(() => res.status(200).send("Task deleted"))
+        .catch(err => res.status(404).send(err));
+    }
   });
 };
